msm: ipa: align to latest kernel version
Make changes to align to the latest kernel version.
Fixes to compilation errors and forbidden warnings.

+++ drivers/platform/msm/ipa/ipa_v3/ipahal/ipahal_hw_stats.c
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-2.0-only
 /*
- * Copyright (c) 2017-2019, The Linux Foundation. All rights reserved.
+ * Copyright (c) 2017-2021, The Linux Foundation. All rights reserved.
  */
 
 #include "ipahal.h"
@@ -480,7 +480,7 @@ static struct ipahal_stats_init_pyld *ipahal_generate_init_pyld_flt_rt_v4_5(
 	void *params, bool is_atomic_ctx)
 {
 	struct ipahal_stats_init_pyld *pyld;
-	int num = (int)(params);
+	long num = (long)(params); /* params is treated as a pointer. */
 
 	if (num > IPA_MAX_FLT_RT_CNT_INDEX ||
 		num <= 0) {
